Separate `withStore` HOC and add SizesProvider
